WIP: Add LiveFeed v5 client
Based on the search client, the code is more or less a copy-paste
of the search client, modified to make it work with the new
livefeed api. The main difference between the two:

There is no need for a Query class for LiveFeed. It has one
endpoint with three input parameters which is easier to just
handle directly in the Client. (In Search, it is the Client
object that handles the request anyway.)