Fix tests
There was a conflict between the classes in the LiveFeed and Search
namespace. This is because we run "include Twingly::Search" in all
spec files. The specs for LiveFeed then used the classes in the
Search namespace and vice versa.

This commit doesn't fix the underlying problem, but now all specs
passes.