Make sure UTC time is used in the specs
The request url used in the specs contained a timestamp which was
in the local time zone. Because of this, running the specs on Travis
(which has UTC) failed because VCR couldn't find a matching request.

The timestamp sent in the request is now in UTC. This only matters
in the specs, as the LiveFeed API should be able to handle both UTC
and non-UTC timestamps.