Use latest version of LiveFeed response
Some fields in the reponse from LiveFeed has been changed:

* Added nextTimestamp
* noOfPosts -> numberOfPosts
* indexed   -> indexedAt (ISO 8601)
* published -> publishedAt (ISO 8601)
* reindexed -> reindexedAt (ISO 8601)

Removed the get_data as there is not much point in having it since
the introduction of nextTimestamp.

Also it was a bit confusing because it updated the timestamp before
the call, meaning that Client#timestamp returned the timestamp for
the last request. When using Client#next_result, #timestamp contains
the timestamp for the next request.